@@ -1,22 +1,147 @@
+# ---------------------------------------------------------------------------
+# pyproject.toml — modernised for 2025‑07 according to PEP 621 and Hatch
+# ---------------------------------------------------------------------------
+
 [build-system]
-requires = ["hatchling"]
+requires = [
+    "hatchling>=1.24",
+    "hatch-vcs>=0.3",
+]
 build-backend = "hatchling.build"
 
+# ---------------------------------------------------------------------------
+# Project metadata (PEP 621)
+# ---------------------------------------------------------------------------
 [project]
-name = "flusstools"
-version = "1.1.8"
-authors = [
-  { name="FlussTeam/Hydro-Informatics.com", email="schwindt.water@gmail.com" },
-]
-description = "A Python package to facilitate river analysis."
-readme = "README.md"
-requires-python = ">=3.9"
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: BSD License",
-    "Operating System :: OS Independent",
+name            = "flusstools"
+dynamic         = ["version"]
+description     = "A Python toolbox for river and floodplain analysis."
+readme          = "README.md"
+requires-python = ">=3.10"
+license         = { text = "BSD-3-Clause" }
+authors         = [{ name = "FlussTeam / Hydro‑Informatics", email = "schwindt.water@gmail.com" }]
+keywords = ["rivers", "hydro‑morphology", "geospatial", "gis", "hydraulic‑engineering", "hydroinformatics",
+  "numerical", "simulations", "Python", "QGIS", "schwindt"]
+classifiers     = [
+  "Development Status :: 4 - Beta",
+  "Intended Audience :: Science/Research",
+  "License :: OSI Approved :: BSD License",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3 :: Only",
+  "Programming Language :: Python :: 3.13",
+  "Topic :: Scientific/Engineering :: GIS",
 ]
 
-[project.urls]
-"Homepage" = "https://flusstools.readthedocs.io/"
-"Bug Tracker" = "https://github.com/Ecohydraulics/flusstools/issues"
+# ---------------------------------------------------------------------------
+# Runtime dependencies
+# ---------------------------------------------------------------------------
+[project.dependencies]
+numpy        = ">=2.3"
+scipy        = ">=1.16"
+pandas       = ">=2.3"
+matplotlib   = ">=3.10"
+networkx     = ">=3.5"
+# Geospatial
+geopandas    = ">=1.1"
+shapely      = ">=2.1"
+rasterio     = ">=1.4"
+laspy        = ">=2.6"
+mapclassify  = ">=2.9"
+pyshp        = ">=2.3"
+rasterstats  = ">=0.20"
+h5py         = ">=3.14"
+alphashape   = ">=1.3"
+earthpy      = ">=0.9"
+geojson      = ">=3.2"
+tabulate     = ">=0.9"
+
+# ---------------------------------------------------------------------------
+# Optional extras
+# ---------------------------------------------------------------------------
+[project.optional-dependencies]
+docs  = [
+  "sphinx>=7.3",
+  "myst-parser>=2.0",
+  "sphinx-book-theme>=1.1",
+  "sphinxcontrib-bibtex>=2.6",
+]
+
+test  = [
+  "pytest>=8.0",
+  "pytest-cov>=5.0",
+  "pytest-xdist>=3.5",
+]
+
+style = [
+  "ruff>=0.4",
+  "pre-commit>=3.7",
+]
+
+# ---------------------------------------------------------------------------
+# Hatch build settings
+# ---------------------------------------------------------------------------
+[tool.hatch.build.targets.wheel]
+include = [
+  "src/flusstools/**",
+  "src/flusstools/py.typed",
+  "CHANGELOG.md",
+  "README.md",
+]
+
+[tool.hatch.version]
+source = "vcs"
+
+# ---------------------------------------------------------------------------
+# Ruff configuration (0.4+)
+# ---------------------------------------------------------------------------
+[tool.ruff]
+line-length    = 100
+exclude        = ["docs", "tests/data"]
+target-version = "py313"
+
+[tool.ruff.lint]
+select = ["E", "F", "I", "UP", "NPY"]        # core rules only
+ignore = ["ANN001", "D211", "D213", "D400", "D401", "D415", "F401", "F403", "F405", "E501", "E712", "E721", "E722",
+  "E741", "F841", "EXE002", "RUF007", "UP030", "UP031"]     # ignore import and docstring conflicts
+
+[tool.ruff.lint.per-file-ignores]
+"__init__.py" = ["F401"]
+
+[tool.ruff.format]
+indent-style = "space"
+line-ending  = "auto"
+
+# ---------------------------------------------------------------------------
+# Mypy type checker (optional strict settings)
+# ---------------------------------------------------------------------------
+[tool.mypy]
+python_version        = "3.13"
+strict                = true
+show_error_codes      = true
+ignore_missing_imports = true
+
+# ---------------------------------------------------------------------------
+# Hatch environments
+# ---------------------------------------------------------------------------
+[tool.hatch.envs.style]
+python       = "3.13"
+dependencies = ["ruff>=0.4", "pre-commit>=3.7"]
+post-install = "pre-commit install"
+
+[tool.hatch.envs.test]
+python       = "3.13"
+dependencies = [
+  "pytest>=8.0",
+  "pytest-cov>=5.0",
+  "pytest-xdist>=3.5",
+]
+
+[tool.hatch.envs.docs]
+python       = "3.13"
+dependencies = [
+  "sphinx>=7.3",
+  "myst-parser>=2.0",
+  "sphinx-book-theme>=1.1",
+  "sphinxcontrib-bibtex>=2.6",
+]
+post-install = "sphinx-build -M html docs docs/_build"
