@@ -21,10 +21,10 @@ class LasPoint:
         scale (laspy.file.File().header.scale): Scale of las points relative to the offset (auto-read)
         shapefile_name (str): The name and dicrectorty of a point shapefile where all las-file data is stored
         srs (osr.SpatialReference): The geo-spatial reference imported from ``epsg``
+
     """
 
     def __init__(self, las_file_name, epsg=3857, use_attributes="aciw", overwrite=True):
-
         self.las_file = laspy.file.File(las_file_name, mode="r")
 
         self.attributes = use_attributes
@@ -68,41 +68,50 @@ class LasPoint:
 
         Returns:
             int: ``0`` if successful, otherwise ``-1``
+
         """
         logging.info(" * Creating GeoTIFF DEM %s ..." % target_file_name)
 
-        default_keys = {"src_shp_file_name": self.shapefile_name,
-                        "elevation_field_name": "Elevation",
-                        "interpolate_gap_pixels": False,
-                        "radius1": -1,
-                        "radius2": -1,
-                        "power": 1.0,
-                        "smoothing": 0.0,
-                        "min_points": 0,
-                        "max_points": 0,
-                        }
+        default_keys = {
+            "src_shp_file_name": self.shapefile_name,
+            "elevation_field_name": "Elevation",
+            "interpolate_gap_pixels": False,
+            "radius1": -1,
+            "radius2": -1,
+            "power": 1.0,
+            "smoothing": 0.0,
+            "min_points": 0,
+            "max_points": 0,
+        }
 
-        for k in default_keys.keys():
+        for k in default_keys:
             if kwargs.get(k):
                 default_keys[k] = str(kwargs.get(k))
 
         if not os.path.isfile(default_keys["src_shp_file_name"]):
-            logging.info(" * Need to create a point shapefile first (%s does not exist) ..." % default_keys["src_shp_file_name"])
+            logging.info(
+                " * Need to create a point shapefile first (%s does not exist) ..."
+                % default_keys["src_shp_file_name"]
+            )
             self.export2shp(shapefile_name=default_keys["src_shp_file_name"])
 
         if os.path.isfile(target_file_name) and self.overwrite:
             logging.info("   -- Overwriting %s ..." % target_file_name)
             geo_utils.remove_tif(target_file_name)
 
-        geo_utils.rasterize(default_keys["src_shp_file_name"], target_file_name, pixel_size=pixel_size,
-                            field_name=default_keys["elevation_field_name"],
-                            interpolate_gap_pixels=default_keys["interpolate_gap_pixels"],
-                            power=default_keys["power"],
-                            radius1=default_keys["radius1"],
-                            radius2=default_keys["radius2"],
-                            smoothing=default_keys["smoothing"],
-                            min_points=default_keys["min_points"],
-                            max_points=default_keys["max_points"])
+        geo_utils.rasterize(
+            default_keys["src_shp_file_name"],
+            target_file_name,
+            pixel_size=pixel_size,
+            field_name=default_keys["elevation_field_name"],
+            interpolate_gap_pixels=default_keys["interpolate_gap_pixels"],
+            power=default_keys["power"],
+            radius1=default_keys["radius1"],
+            radius2=default_keys["radius2"],
+            smoothing=default_keys["smoothing"],
+            min_points=default_keys["min_points"],
+            max_points=default_keys["max_points"],
+        )
         logging.info("   -- Done.")
         return 0
 
@@ -112,13 +121,15 @@ class LasPoint:
         Keyword Args:
             shapefile_name (`str`): Optional shapefile name (must end on .shp).
                                         (default: ``'/this/dir/las_file_name.shp'``).
+
         Returns:
             str: ``/path/to/shapefile.shp``, which is a point shapefile created by the function.
+
         """
         if kwargs.get("shapefile_name"):
             self.shapefile_name = kwargs.get("shapefile_name")
         else:
-            self.shapefile_name = os.path.abspath("") + "/{0}.shp".format(self.las_file.filename)
+            self.shapefile_name = os.path.abspath("") + f"/{self.las_file.filename}.shp"
 
         if os.path.isfile(self.shapefile_name) and self.overwrite is False:
             logging.info(" * Using existing shapefile %s." % self.shapefile_name)
@@ -126,15 +137,16 @@ class LasPoint:
 
         self._build_data_frame()
 
-        logging.info(" * Writing geopandas.GeoDataFrame to shapefile (%s) ..." % self.shapefile_name)
+        logging.info(
+            " * Writing geopandas.GeoDataFrame to shapefile (%s) ..." % self.shapefile_name
+        )
         logging.info("   *** this action may take a while (0.5h per 1 million points)***")
         self.gdf.to_file(filename=self.shapefile_name, driver="ESRI Shapefile")
         logging.info("   -- Done.")
         return self.shapefile_name
 
     def get_file_info(self):
-        """ Prints las file information to console."""
-
+        """Prints las file information to console."""
         print("Point data formats in file:")
         for f in self.las_file.point_format:
             print("   -- %s" % f.name)
@@ -143,14 +155,13 @@ class LasPoint:
         print("   -- " + ", ".join(headers))
 
     def _build_data_frame(self):
-        """ Builds the geopandas GeoDataFrame - auto-runs ``self._parse_attributes``."""
+        """Builds the geopandas GeoDataFrame - auto-runs ``self._parse_attributes``."""
         point_dict = self._parse_attributes()
         # for attr in self.pts_description:
         #     if not re.search("[x-z]", attr):
         #         point_dict.update({attr: pts_df[attr]})
         logging.info(" * Building geopandas.GeoDataFrame ...")
-        self.gdf = geopandas.GeoDataFrame(pd.DataFrame(point_dict),
-                                          crs="EPSG:%04i" % self.epsg)
+        self.gdf = geopandas.GeoDataFrame(pd.DataFrame(point_dict), crs="EPSG:%04i" % self.epsg)
         logging.info("   -- Done.")
 
     def _get_xyz_array(self):
@@ -158,11 +169,12 @@ class LasPoint:
 
         Returns:
             ndarray: The DEM information extracted from the las file.
+
         """
-        pts = self.las_file.points['point'].copy().view(np.recarray)
+        pts = self.las_file.points["point"].copy().view(np.recarray)
 
         # read and transform data (from raw - fast than las_file.x)
-        dem_array = np.empty((3, len(pts)), dtype=np.float64)
+        dem_array = np.empty((3, len(pts)), dtype=float)
         dem_array[0, :] = pts.X * self.scale[0] + self.offset[0]
         dem_array[1, :] = pts.Y * self.scale[1] + self.offset[1]
         dem_array[2, :] = pts.Z * self.scale[2] + self.offset[2]
@@ -171,7 +183,6 @@ class LasPoint:
 
     def _parse_attributes(self):
         """Parses attributes and append entries to point list."""
-
         logging.info(" * Extracting transformed point coordinates ...")
         dem = self._get_xyz_array()
         point_dict = {"geometry": geopandas.points_from_xy(x=dem[0], y=dem[1], z=dem[2])}
@@ -184,12 +195,13 @@ class LasPoint:
                 point_dict.update({wattr[attr]: self.las_file.__getattribute__(pattr[attr])})
                 logging.info("   -- added %s" % wattr[attr])
             except AttributeError:
-                logging.error("Non-existing attribute %s. Valid attributes are: %s" % (str(attr), dict2str(wattr)))
+                logging.error(
+                    "Non-existing attribute %s. Valid attributes are: %s"
+                    % (str(attr), dict2str(wattr))
+                )
             except KeyError:
-                logging.error("Non-existing las-file key %s - valid are: " % str(attr) + ", ".join(dir(self.las_file)))
+                logging.error(
+                    "Non-existing las-file key %s - valid are: " % str(attr)
+                    + ", ".join(dir(self.las_file))
+                )
         return point_dict
-
-
-
-
-
