@@ -1,13 +1,14 @@
-import sys
 import os
+import sys
+
 sys.path.insert(0, os.path.dirname(__file__))
 
 try:
-    from helpers import *
-    import lidartools
+    import bedanalyst
     import fuzzycorr
     import geotools
-    import bedanalyst
+    import lidartools
+    from helpers import *
 except ModuleNotFoundError:
     print("Failed to initialize FlussTools - consider re-installation")
 
