@@ -1,10 +1,11 @@
 from .raster_mgmt import *
 from .shp_mgmt import *
+
 gdal.UseExceptions()
 
 
 def coords2offset(geo_transform, x_coord, y_coord):
-    """ Returns x-y pixel offset (inverse of the ``offset2coords`` function).
+    """Returns x-y pixel offset (inverse of the ``offset2coords`` function).
 
     Args:
         geo_transform: osgeo.gdal.Dataset.GetGeoTransform() object
@@ -13,6 +14,7 @@ def coords2offset(geo_transform, x_coord, y_coord):
 
     Returns:
         tuple: Number of pixels ``(offset_x, offset_y)``,  both ``int`` .
+
     """
     try:
         origin_x = geo_transform[0]
@@ -20,16 +22,14 @@ def coords2offset(geo_transform, x_coord, y_coord):
         pixel_width = geo_transform[1]
         pixel_height = geo_transform[5]
     except IndexError:
-        logging.error("Invalid geo_transform object (%s)." %
-                      str(geo_transform))
+        logging.error("Invalid geo_transform object (%s)." % str(geo_transform))
         return None
 
     try:
         offset_x = int((x_coord - origin_x) / pixel_width)
         offset_y = int((y_coord - origin_y) / pixel_height)
     except ValueError:
-        logging.error(
-            "geo_transform tuple contains non-numeric data: %s" % str(geo_transform))
+        logging.error("geo_transform tuple contains non-numeric data: %s" % str(geo_transform))
         return None
     return offset_x, offset_y
 
@@ -43,11 +43,12 @@ def get_layer(dataset, band_number=1):
 
     Returns:
         dict: ``{"type": raster`` or ``vector`` or ``"None", layer":`` if raster: ``raster_band``, if vector: ``GetLayer()``, else: ``None}``
+
     """
     if verify_dataset(dataset) == "raster":
         return {"type": "raster", "layer": dataset.GetRasterBand(band_number)}
     if verify_dataset(dataset) == "vector":
-        return {"type": "vector", "layer":  dataset.GetLayer()}
+        return {"type": "vector", "layer": dataset.GetLayer()}
     return {"type": "None", "layer": None}
 
 
@@ -61,6 +62,7 @@ def offset2coords(geo_transform, offset_x, offset_y):
 
     Returns:
         tuple: Two ``float`` numbers of x-y-coordinates ``(x_coord, y_coord)``.
+
     """
     try:
         origin_x = geo_transform[0]
@@ -68,16 +70,14 @@ def offset2coords(geo_transform, offset_x, offset_y):
         pixel_width = geo_transform[1]
         pixel_height = geo_transform[5]
     except IndexError:
-        logging.error("Invalid geo_transform object (%s)." %
-                      str(geo_transform))
+        logging.error("Invalid geo_transform object (%s)." % str(geo_transform))
         return None
 
     try:
         coord_x = origin_x + pixel_width * (offset_x + 0.5)
         coord_y = origin_y + pixel_height * (offset_y + 0.5)
     except ValueError:
-        logging.error(
-            "geo_transform tuple contains non-numeric data: %s" % str(geo_transform))
+        logging.error("geo_transform tuple contains non-numeric data: %s" % str(geo_transform))
         return None
     return coord_x, coord_y
 
@@ -90,6 +90,7 @@ def verify_dataset(dataset):
 
     Returns:
         str: Either "unknown", "raster", or "vector".
+
     """
     # Check the contents of an osgeo.gdal.Dataset
     try:
@@ -107,8 +108,7 @@ def verify_dataset(dataset):
     try:
         if dataset.GetLayerCount() > 0:
             return "vector"
-        else:
-            return "empty"
+        return "empty"
     except AttributeError:
         logging.error("%s is not an osgeo.gdal.Dataset object." % str(dataset))
         return None
