@@ -1,13 +1,16 @@
 """Global variables"""
+
 from var_config import *
 
 
 def cache(fun):
     """Makes a function running in a temporary ``__cache__`` sub-folder to enable deleting temporary trash files."""
+
     def wrapper(*args, **kwargs):
         check_cache()
         fun(*args, **kwargs)
         remove_directory(cache_folder)
+
     wrapper.__doc__ = fun.__doc__
     return wrapper
 
@@ -24,7 +27,6 @@ def check_if_file_in_use(filepath):
     """Checks if a file or list of files is in use by another process.
     If the file cannot be opened or there is an associated .lock file, it throws an exception.
     """
-
     if type(filepath) == list:
         for f in filepath:
             check_if_file_in_use(f)
@@ -35,20 +37,26 @@ def check_if_file_in_use(filepath):
         try:
             buffer_size = 8
             # Opening file in append mode and read the first 8 characters.
-            file_object = open(filepath, 'a', buffer_size)
+            file_object = open(filepath, "a", buffer_size)
             if file_object:
                 for filename in os.listdir(os.path.dirname(filepath)):
-                    if filename.startswith(os.path.basename(filepath)) and filename.endswith('.lock'):
+                    if filename.startswith(os.path.basename(filepath)) and filename.endswith(
+                        ".lock"
+                    ):
                         logging.error(
-                            '%s is open in another program. Close the file and try again.' % filepath)
+                            "%s is open in another program. Close the file and try again."
+                            % filepath
+                        )
                         raise Exception(
-                            '%s is open in another program. Close the file and try again.' % filepath)
+                            "%s is open in another program. Close the file and try again."
+                            % filepath
+                        )
 
-        except IOError:
-            logging.error(
-                '%s is open in another program. Close the file and try again.' % filepath)
+        except OSError:
+            logging.error("%s is open in another program. Close the file and try again." % filepath)
             raise Exception(
-                '%s is open in another program. Close the file and try again.' % filepath)
+                "%s is open in another program. Close the file and try again." % filepath
+            )
 
         finally:
             if file_object:
@@ -63,31 +71,37 @@ def err_info(func):
             func(*args, **kwargs)
         except Exception as e:
             logging.info(e)
-            messagebox.showerror('Error', e)
+            messagebox.showerror("Error", e)
+
     return wrapper
 
 
-def get_file_names(directory, prefix='', suffix='', nesting=True):
-    """
-    Returns list of all files in directory
+def get_file_names(directory, prefix="", suffix="", nesting=True):
+    """Returns list of all files in directory
 
     Args:
         directory (str): the directory of interest
         prefix (str): if provided, files returned must start with this
         suffix (str): if provided, files returned must end with this
         nesting (bool): if True, looks in all subdirectories of dir. If false, only looks at top-level.
+
     """
     l = []
     for path, subdirs, files in os.walk(directory):
         for name in files:
-            if name.startswith(prefix) and name.endswith(suffix) and (nesting or (path == directory)):
+            if (
+                name.startswith(prefix)
+                and name.endswith(suffix)
+                and (nesting or (path == directory))
+            ):
                 l.append(os.path.join(path, name))
     return l
 
 
-def lookup_value(df, value, src_column_name, lookup_column_name, mode="single_column", src_column_name2=""):
-    """
-    !!!FUNCTION NOT YET DEBUGGED!!!
+def lookup_value(
+    df, value, src_column_name, lookup_column_name, mode="single_column", src_column_name2=""
+):
+    """!!!FUNCTION NOT YET DEBUGGED!!!
     lookup a value from a pandas DataFrame similar to VLOOKUP.
     The function can read either the first value that is within a threshold value frame of
      a single column, or two columns (mode argument).
@@ -102,6 +116,7 @@ def lookup_value(df, value, src_column_name, lookup_column_name, mode="single_co
 
     Returns:
         Value of the ``lookup_column_name`` corresponding to ``value`` in ``src_column_name``.
+
     """
     if mode == "single_column":
         match = (df[src_column_name] <= value) & (df[src_column_name] > value)
@@ -122,6 +137,7 @@ def remove_directory(directory):
 
     Returns:
         None: Deletes directory.
+
     """
     try:
         for root, dirs, files in os.walk(directory):
