@@ -1,3 +1,3 @@
+from .hdf_analyst import *
 from .laspy_main import *
 from .lastools_core import *
-from .hdf_analyst import *
